Make the return value of minimumDataSize() signed.
Done to be able to report an error following the idea of
calculateDataSize(...). Fix stream operator implementation.
Adjust auto-test. Adjust minimumDataSize(...) to be able
to handle unknow minimum sizes for certain function codes.

+++ src/serialbus/qmodbuspdu.h
@@ -194,7 +194,7 @@ public:
         : QModbusPdu(code, newData)
     {}
 
-    Q_SERIALBUS_EXPORT static quint8 minimumDataSize(FunctionCode code);
+    Q_SERIALBUS_EXPORT static int minimumDataSize(FunctionCode code);
     Q_SERIALBUS_EXPORT static int calculateDataSize(FunctionCode code, const QByteArray &data);
 };
 Q_SERIALBUS_EXPORT QDataStream &operator>>(QDataStream &stream, QModbusRequest &pdu);
@@ -214,7 +214,7 @@ public:
         : QModbusPdu(code, newData)
     {}
 
-    Q_SERIALBUS_EXPORT static quint8 minimumDataSize(FunctionCode code);
+    Q_SERIALBUS_EXPORT static int minimumDataSize(FunctionCode code);
     Q_SERIALBUS_EXPORT static int calculateDataSize(FunctionCode code, const QByteArray &data);
 };
